flask-backend/requirement.txt
Flask == 2.3.2
Flask-SQLAlchemy == 3.0.3
Flask-Migrate == 4.0.4
Flask-JWT-Extended == 4.5.2
gunicorn
python-dotenv
flask_cors
minio
flask_login
flask_admin
schedule